parsers: moved to_html out of parse_webvtt_to_div
